+++ packages/react-widgets/src/widgets/HistogramWidget.js
@@ -126,7 +126,7 @@ function HistogramWidget({
     dataSource,
     id,
     column,
-    type: FilterTypes.CLOSED_OPEN
+    type: FilterTypes.BETWEEN
   });
 
   const selectedBars = useMemo(() => {
@@ -138,7 +138,7 @@ function HistogramWidget({
           return ticks.length;
         } else {
           const idx = ticks.indexOf(from);
-          return idx !== -1 ? idx + 1 : null;
+          return idx !== -1 ? idx + 1 : 0;
         }
       })
       .filter((v) => v !== null);
@@ -148,8 +148,8 @@ function HistogramWidget({
     (selectedBars) => {
       if (selectedBars?.length) {
         const thresholds = selectedBars.map((i) => {
-          let left = ticks[i - 1];
-          let right = ticks.length !== i ? ticks[i] : undefined;
+          let left = ticks[i - 1] || min;
+          let right = ticks.length !== i ? ticks[i] : max;
 
           return [left, right];
         });
@@ -157,7 +157,7 @@ function HistogramWidget({
           addFilter({
             id: dataSource,
             column,
-            type: FilterTypes.CLOSED_OPEN,
+            type: FilterTypes.BETWEEN,
             values: thresholds,
             owner: id
           })
@@ -172,7 +172,7 @@ function HistogramWidget({
         );
       }
     },
-    [column, dataSource, id, dispatch, ticks]
+    [column, dataSource, id, dispatch, ticks, min, max]
   );
 
   return (